dags/autohrv/requirements.txt
alembic==1.14.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
fastapi==0.112.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
lxml==5.2.2
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-editor==1.0.4
requests==2.32.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
